requirement.txt
# --- Core Frameworks ---
apache-airflow==2.10.5
dbt-core==1.10.13
dbt-postgres==1.9.1

# --- Data Manipulation & Analysis ---
pandas
numpy
pyarrow

# --- Database & Messaging ---
psycopg2-binary
pika==1.1.0

# --- Machine Learning ---
scikit-learn
joblib

# --- Visualization (Optional - if used in notebooks) ---
plotly
matplotlib